Added functionality int he update route to create timestamps [PAR-101].

+++ db.js
@@ -450,36 +450,55 @@ var loans = (function () {
       updateQuery = mysql.format(updateQuery, updateInserts);
 
       var currentDate = moment().format("YYYY-MM-DD HH:mm:ss");
-      var endDate = moment().add(1, "months");
-      endDate = endDate.format("YYYY-MM-DD HH:mm:ss");
-      console.log(currentDate);
-      console.log(endDate);
 
-      return pool.query(updateQuery).then((updateResult) => {
-        if (loan.status === "accepted") {
-          let deleteQuery =
-            "DELETE FROM loans WHERE status = 'pending' AND book_id = ?";
-          let deleteInserts = [loan.book_id];
-          deleteQuery = mysql.format(deleteQuery, deleteInserts);
+      var updateTimeQuery = "UPDATE loans SET ";
+      // After createing an update query, fetching the current time, and creating a starting updateTime query,
+      // check to see if the status is anything but "accepted." If so, set the time query to update its respective field.
 
-          pool.query(deleteQuery);
+      if (loan.status === "loaned") {
+        updateTimeQuery = updateTimeQuery + "loan_start_date = ? WHERE id = ?";
+        let updateTimeInserts = [currentDate, loan.id];
+        updateTimeQuery = mysql.format(updateTimeQuery, updateTimeInserts);
 
-          let updateTimeQuery =
-            "UPDATE loans SET accept_date = ?, loan_start_date = ?, " +
-            "loan_end_date = ?, return_date = ? WHERE id = ?";
-          let updateTimeInserts = [
-            currentDate,
-            currentDate,
-            endDate,
-            endDate,
-            loan.id,
-          ];
-          updateTimeQuery = mysql.format(updateTimeQuery, updateTimeInserts);
+        pool.query(updateTimeQuery);
+      } else if (loan.status === "returned") {
+        updateTimeQuery = updateTimeQuery + "return_date = ? WHERE id = ?";
+        let updateTimeInserts = [currentDate, loan.id];
+        updateTimeQuery = mysql.format(updateTimeQuery, updateTimeInserts);
 
-          pool.query(updateTimeQuery);
-        }
-        return updateResult;
-      });
+        pool.query(updateTimeQuery);
+      } else if (loan.status === "canceled" || loan.status === "declined") {
+        var queuedForDeletion = 1; // Sets a delete flag
+        var deleteQuery = "DELETE FROM loans WHERE id = ?";
+        var inserts = [loan.id];
+        deleteQuery = mysql.format(deleteQuery, inserts);
+      }
+
+      if (queuedForDeletion === 1) {
+        // If the goal is to delete the loan, we don't want to update it.
+        // As a result, this only updates the loan if the delete flag has been reached.
+        return pool.query(deleteQuery);
+      } else {
+        // If we have not reached the delete flag, then query the initial updateQuery from the top,
+        // and check to see if the status === "accepted."
+        return pool.query(updateQuery).then((updateResult) => {
+          if (loan.status === "accepted") {
+            let accDeleteQuery =
+              "DELETE FROM loans WHERE status = 'pending' AND book_id = ?";
+            let deleteInserts = [loan.book_id];
+            accDeleteQuery = mysql.format(accDeleteQuery, deleteInserts);
+
+            pool.query(accDeleteQuery);
+
+            updateTimeQuery = updateTimeQuery + "accept_date = ? WHERE id = ?";
+            let updateTimeInserts = [currentDate, loan.id];
+            updateTimeQuery = mysql.format(updateTimeQuery, updateTimeInserts);
+
+            pool.query(updateTimeQuery);
+          }
+          return updateResult;
+        });
+      }
     },
 
     deleteLoan: function (loanId) {
